Nonstopflights query doc update (#283)
* Update public-api.md

Updated the 'query' sample as it does not work with a copy and paste directly from the docs. Not 100% sure this update is the intended outcome wanted in the query, but it will at least bring light to the issue.

* Update public-api.md

I'm not 100% sure if this is the intended query, but the example on the site does not work when copy and pasted directly. If not correct, it can at least serve to bring light to the issue. Thanks!